3.8 hecho pero no funciona, mañana será otro día. 22/01

+++ src/main/java/org/iesvdm/dao/ComercialDAOImpl.java
@@ -44,7 +44,7 @@ public class ComercialDAOImpl implements ComercialDAO {
 			ps.setString(idx++, comercial.getNombre());
 			ps.setString(idx++, comercial.getApellido1());
 			ps.setString(idx++, comercial.getApellido2());
-			ps.setDouble(idx, comercial.getComision());
+			ps.setBigDecimal(idx, comercial.getComision());
 			return ps;
 		},keyHolder);
 		
@@ -72,7 +72,7 @@ public class ComercialDAOImpl implements ComercialDAO {
                 							  rs.getString("nombre"), 
                 							  rs.getString("apellido1"),
                 							  rs.getString("apellido2"), 
-                							  rs.getFloat("comisión"))      						 	
+                							  rs.getBigDecimal("comisión"))      						 	
         );
 		
 		log.info("Devueltos {} registros.", listComercial.size());
@@ -88,7 +88,7 @@ public class ComercialDAOImpl implements ComercialDAO {
             						 							rs.getString("nombre"),
             						 							rs.getString("apellido1"),
             						 							rs.getString("apellido2"),
-            						 							rs.getFloat("comisión"))
+            						 							rs.getBigDecimal("comisión"))
 								, id
 								);
