+++ src/main/java/org/iesvdm/dto/ComercialDTO.java
@@ -1,11 +1,14 @@
 package org.iesvdm.dto;
 
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.Optional;
 
 import org.iesvdm.modelo.Cliente;
 import org.iesvdm.modelo.Comercial;
 
+import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.Size;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 
@@ -29,7 +32,7 @@ public class ComercialDTO {
 	private String nombre;
 	private String apellido1;
 	private String apellido2;
-	private float comision;
+	private BigDecimal comision;
 	
 	private float pedidomin;
 	private float pedidomax;
